Update Sales Prediction.cpp

+++ Sales Prediction.cpp
@@ -5,9 +5,9 @@ using namespace std;
 int main()
 
 {
-	double percentageofTotalSales = .58;
-	double salesThisYear = 8.6;
-	double result = salesThisYear * percentageofTotalSales;
+	float percentageofTotalSales = .58;
+	float salesThisYear = 8.6;
+	float result = salesThisYear * percentageofTotalSales;
 	cout<<"The Total Sales That The East Cost Divison Made This Year Was $" <<result<< " million. "<<endl;
 	getch();
 	return 0;
